archive/backend-legacy-20251015/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
httpx==0.24.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2
pypdf==4.2.0
rank-bm25==0.2.2
PyYAML==6.0.2
